Moved the DbContext and User, renamed namespaces
