+++ HolidayDestinations/HolidayDestinations.Web/Account/Register.aspx.cs
@@ -1,14 +1,13 @@
-﻿using System;
-using System.Linq;
-using System.Web;
-using System.Web.UI;
-using Microsoft.AspNet.Identity;
-using Microsoft.AspNet.Identity.Owin;
-using Owin;
-using HolidayDestinations.Web.Models;
-
-namespace HolidayDestinations.Web.Account
+﻿namespace HolidayDestinations.Web.Account
 {
+    using System;
+    using System.Linq;
+    using System.Web;
+    using System.Web.UI;
+    using Microsoft.AspNet.Identity;
+    using Microsoft.AspNet.Identity.Owin;
+    using Owin;
+    using Models;
     public partial class Register : Page
     {
         protected void CreateUser_Click(object sender, EventArgs e)
