+++ HolidayDestinations/HolidayDestinations.Web/Account/ResetPassword.aspx.cs
@@ -5,7 +5,6 @@ using System.Web.UI;
 using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.Owin;
 using Owin;
-using HolidayDestinations.Web.Models;
 
 namespace HolidayDestinations.Web.Account
 {
