+++ HolidayDestinations/HolidayDestinations.Web/App_Start/Startup.Auth.cs
@@ -1,16 +1,16 @@
-﻿using System;
-using Microsoft.AspNet.Identity;
-using Microsoft.AspNet.Identity.EntityFramework;
-using Microsoft.AspNet.Identity.Owin;
-using Microsoft.Owin;
-using Microsoft.Owin.Security.Cookies;
-using Microsoft.Owin.Security.DataProtection;
-using Microsoft.Owin.Security.Google;
-using Owin;
-using HolidayDestinations.Web.Models;
-
-namespace HolidayDestinations.Web
+﻿namespace HolidayDestinations.Web
 {
+    using System;
+    using Microsoft.AspNet.Identity;
+    using Microsoft.AspNet.Identity.EntityFramework;
+    using Microsoft.AspNet.Identity.Owin;
+    using Microsoft.Owin;
+    using Microsoft.Owin.Security.Cookies;
+    using Microsoft.Owin.Security.DataProtection;
+    using Microsoft.Owin.Security.Google;
+    using Owin;
+    using Data;
+    using Models;
     public partial class Startup {
 
         // For more information on configuring authentication, please visit http://go.microsoft.com/fwlink/?LinkId=301883
