+++ HolidayDestinations/HolidayDestinations.Web/HolidayDestinations.Web.csproj
@@ -427,7 +427,18 @@
     </Compile>
   </ItemGroup>
   <ItemGroup>
-    <Folder Include="App_Data\" />
+    <ProjectReference Include="..\HolidayDestinations.Data\HolidayDestinations.Data.csproj">
+      <Project>{6c433ed1-c9e4-4464-af81-31355738d294}</Project>
+      <Name>HolidayDestinations.Data</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\HolidayDestinations.Models\HolidayDestinations.Models.csproj">
+      <Project>{251ec757-4740-4f06-bd15-9c0712e43f47}</Project>
+      <Name>HolidayDestinations.Models</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\HolidayDestinations.Services\HolidayDestinations.Services.csproj">
+      <Project>{f9f0e3d9-093f-4f36-8e2d-e06c15c26733}</Project>
+      <Name>HolidayDestinations.Services</Name>
+    </ProjectReference>
   </ItemGroup>
   <PropertyGroup>
     <VisualStudioVersion Condition="'$(VisualStudioVersion)' == ''">10.0</VisualStudioVersion>
