+++ HolidayDestinations/HolidayDestinations.Web/Models/IdentityModels.cs
@@ -1,49 +1,8 @@
-﻿using System;
-using System.Security.Claims;
-using System.Threading.Tasks;
-using System.Web;
-using Microsoft.AspNet.Identity;
-using Microsoft.AspNet.Identity.EntityFramework;
-using Microsoft.AspNet.Identity.Owin;
-using Microsoft.Owin.Security;
-using HolidayDestinations.Web.Models;
-
-namespace HolidayDestinations.Web.Models
+﻿namespace HolidayDestinations.Web
 {
-    // You can add User data for the user by adding more properties to your User class, please visit http://go.microsoft.com/fwlink/?LinkID=317594 to learn more.
-    public class User : IdentityUser
-    {
-        public ClaimsIdentity GenerateUserIdentity(ApplicationUserManager manager)
-        {
-            // Note the authenticationType must match the one defined in CookieAuthenticationOptions.AuthenticationType
-            var userIdentity = manager.CreateIdentity(this, DefaultAuthenticationTypes.ApplicationCookie);
-            // Add custom user claims here
-            return userIdentity;
-        }
+    using System;
+    using System.Web;
 
-        public Task<ClaimsIdentity> GenerateUserIdentityAsync(ApplicationUserManager manager)
-        {
-            return Task.FromResult(GenerateUserIdentity(manager));
-        }
-    }
-
-    public class HolidayDesitantionsDbContext : IdentityDbContext<User>
-    {
-        public HolidayDesitantionsDbContext()
-            : base("DefaultConnection", throwIfV1Schema: false)
-        {
-        }
-
-        public static HolidayDesitantionsDbContext Create()
-        {
-            return new HolidayDesitantionsDbContext();
-        }
-    }
-}
-
-#region Helpers
-namespace HolidayDestinations.Web
-{
     public static class IdentityHelper
     {
         // Used for XSRF when linking external logins
@@ -97,4 +56,3 @@ namespace HolidayDestinations.Web
         }
     }
 }
-#endregion
